+++ ebean-core/src/main/java/io/ebeaninternal/server/query/SqlTreeLoadBean.java
@@ -28,6 +28,7 @@ class SqlTreeLoadBean implements SqlTreeLoad {
   final boolean readId;
   private final boolean readIdNormal;
   private final boolean disableLazyLoad;
+  private final boolean readOnlyNoIntercept;
   private final InheritInfo inheritInfo;
   final String prefix;
   private final Map<String, String> pathMap;
@@ -49,6 +50,7 @@ class SqlTreeLoadBean implements SqlTreeLoad {
     this.readId = node.readId;
     this.readIdNormal = readId && !temporalVersions;
     this.disableLazyLoad = node.disableLazyLoad;
+    this.readOnlyNoIntercept = disableLazyLoad && node.readOnly;
     this.partialObject = node.partialObject;
     this.properties = node.properties;
     this.pathMap = node.pathMap;
@@ -145,7 +147,7 @@ class SqlTreeLoadBean implements SqlTreeLoad {
 
     void initBeanType() throws SQLException {
       localDesc = desc;
-      localBean = desc.createEntityBean();
+      localBean = desc.createEntityBean2(readOnlyNoIntercept);
       localIdBinder = idBinder;
     }
