+++ ebean-core/src/main/java/io/ebeaninternal/server/query/SqlTreeNodeRoot.java
@@ -20,9 +20,9 @@ final class SqlTreeNodeRoot extends SqlTreeNodeBean {
    * Specify for SqlSelect to include an Id property or not.
    */
   SqlTreeNodeRoot(STreeType desc, SqlTreeProperties props, List<SqlTreeNode> myList, boolean withId, TableJoin includeJoin,
-                  STreePropertyAssocMany many, SpiQuery.TemporalMode temporalMode, boolean disableLazyLoad, boolean sqlDistinct, String baseTable) {
+                  STreePropertyAssocMany many, SpiQuery.TemporalMode temporalMode, boolean disableLazyLoad, boolean readOnly, boolean sqlDistinct, String baseTable) {
 
-    super(desc, props, myList, withId, many, temporalMode, disableLazyLoad);
+    super(desc, props, myList, withId, many, temporalMode, disableLazyLoad, readOnly);
     this.includeJoin = includeJoin;
     this.sqlDistinct = sqlDistinct;
     this.baseTable = baseTable;
