+++ ebean-core/src/test/resources/ebean.mf
@@ -1,0 +1,1 @@
+synthetic: false
