+++ ebean-externalmapping-xml/pom.xml
@@ -98,7 +98,7 @@
         <configuration>
           <tiles>
             <!-- other tiles ... -->
-            <tile>io.ebean.tile:enhancement:12.15.0</tile>
+            <tile>io.ebean.tile:enhancement:13.0.1-RC1</tile>
           </tiles>
         </configuration>
       </plugin>
