+++ ebean-externalmapping-xml/src/test/resources/META-INF/ebean-version.mf
