+++ ebean-kotlin/pom.xml
@@ -14,7 +14,7 @@
   <properties>
     <maven.compiler.source>1.8</maven.compiler.source>
     <maven.compiler.target>1.8</maven.compiler.target>
-    <ebean-maven-plugin.version>12.14.1</ebean-maven-plugin.version>
+<!--    <ebean-maven-plugin.version>13.0.1-SNAPSHOT</ebean-maven-plugin.version>-->
     <kotlin.version>1.6.0</kotlin.version>
   </properties>
 
@@ -93,7 +93,7 @@
         <extensions>true</extensions>
         <configuration>
           <tiles>
-            <tile>io.ebean.tile:enhancement:13.0.0</tile>
+            <tile>io.ebean.tile:enhancement:13.0.1-RC1</tile>
           </tiles>
         </configuration>
       </plugin>
