+++ pom.xml
@@ -46,8 +46,8 @@
     <ebean-migration.version>13.0.0</ebean-migration.version>
     <ebean-test-docker.version>4.5</ebean-test-docker.version>
     <ebean-datasource.version>7.5</ebean-datasource.version>
-    <ebean-agent.version>13.0.0</ebean-agent.version>
-    <ebean-maven-plugin.version>13.0.0</ebean-maven-plugin.version>
+    <ebean-agent.version>13.0.1-RC1</ebean-agent.version>
+    <ebean-maven-plugin.version>13.0.1-RC1</ebean-maven-plugin.version>
     <surefire.useModulePath>false</surefire.useModulePath>
   </properties>
